Elimina la clase Consolidado y separa funciones que no correspondían a clase View. Crea la clase Snies Controller
