--- src/consolidado.py
@@ -1,4 +1,0 @@
-
-class Consolidado:
-    def __init__(self):
-        self.consolidado = {}
